king/telegram-bot/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=21.0
httpx>=0.24.0

